requirements.txt
click==6.7
Flask==0.12
itsdangerous==0.24
Jinja2==2.9.4
MarkupSafe==0.23
python-magic==0.4.12
requests==2.12.4
Werkzeug==0.11.15
